+++ libafl_frida/src/coverage_rt.rs
@@ -1,5 +1,6 @@
 //! Functionality regarding binary-only coverage collection.
 use dynasmrt::{dynasm, DynasmApi, DynasmLabelApi};
+use rangemap::RangeMap;
 use std::ffi::c_void;
 
 #[cfg(target_arch = "x86_64")]
@@ -9,7 +10,7 @@ use frida_gum::instruction_writer::{Aarch64Register, IndexMode};
 
 use frida_gum::{instruction_writer::InstructionWriter, stalker::StalkerOutput};
 
-use crate::helper::{FridaRuntime, FridaInstrumentationHelper};
+use crate::helper::FridaRuntime;
 
 /// (Default) map size for frida coverage reporting
 pub const MAP_SIZE: usize = 64 * 1024;
@@ -31,15 +32,26 @@ impl Default for CoverageRuntime {
 
 impl FridaRuntime for CoverageRuntime {
     /// Initialize the coverage runtime
-    fn init(&mut self, _gum: &frida_gum::Gum, _helper: &FridaInstrumentationHelper, _modules_to_instrument: &[&str]) {
+    fn init(
+        &mut self,
+        _gum: &frida_gum::Gum,
+        ranges: &RangeMap<usize, (u16, String)>,
+        _modules_to_instrument: &[&str],
+    ) {
         self.generate_maybe_log_blob();
     }
 
-    fn pre_exec<I: libafl::inputs::Input + libafl::inputs::HasTargetBytes>(&mut self, _input: &I, _helper: &FridaInstrumentationHelper) -> Result<(), libafl::Error> {
+    fn pre_exec<I: libafl::inputs::Input + libafl::inputs::HasTargetBytes>(
+        &mut self,
+        _input: &I,
+    ) -> Result<(), libafl::Error> {
         Ok(())
     }
 
-    fn post_exec<I: libafl::inputs::Input + libafl::inputs::HasTargetBytes>(&mut self, _input: &I, _helper: &FridaInstrumentationHelper) -> Result<(), libafl::Error> {
+    fn post_exec<I: libafl::inputs::Input + libafl::inputs::HasTargetBytes>(
+        &mut self,
+        _input: &I,
+    ) -> Result<(), libafl::Error> {
         Ok(())
     }
 }
