+++ libafl_frida/src/drcov_rt.rs
@@ -1,11 +1,11 @@
 //! Generates `DrCov` traces
+use crate::helper::FridaRuntime;
 use ahash::AHasher;
 use libafl::{
     inputs::{HasTargetBytes, Input},
     Error,
 };
 use libafl_targets::drcov::{DrCovBasicBlock, DrCovWriter};
-use crate::helper::{FridaRuntime, FridaInstrumentationHelper};
 use rangemap::RangeMap;
 use std::hash::Hasher;
 
@@ -20,20 +20,25 @@ pub struct DrCovRuntime {
 
 impl FridaRuntime for DrCovRuntime {
     /// initializes this runtime wiith the given `ranges`
-    fn init(&mut self, _gum: &frida_gum::Gum, helper: &FridaInstrumentationHelper, _modules_to_instrument: &[&str]) {
-        self.ranges = helper.ranges().clone();
+    fn init(
+        &mut self,
+        _gum: &frida_gum::Gum,
+        ranges: &RangeMap<usize, (u16, String)>,
+        _modules_to_instrument: &[&str],
+    ) {
+        self.ranges = ranges.clone();
         std::fs::create_dir_all("./coverage")
             .expect("failed to create directory for coverage files");
     }
 
     /// Called before execution, does nothing
-    fn pre_exec<I: Input + HasTargetBytes>(&mut self, _input: &I, _helper: &FridaInstrumentationHelper) -> Result<(), Error> {
+    fn pre_exec<I: Input + HasTargetBytes>(&mut self, _input: &I) -> Result<(), Error> {
         Ok(())
     }
 
     /// Called after execution, writes the trace to a unique `DrCov` file for this trace
     /// into `./coverage/<trace_hash>.drcov`
-    fn post_exec<I: Input + HasTargetBytes>(&mut self, input: &I, _helper: &FridaInstrumentationHelper) -> Result<(), Error> {
+    fn post_exec<I: Input + HasTargetBytes>(&mut self, input: &I) -> Result<(), Error> {
         let mut hasher = AHasher::new_with_keys(0, 0);
         hasher.write(input.target_bytes().as_slice());
