+++ libafl_frida/src/executor.rs
@@ -21,7 +21,7 @@ use crate::asan::errors::ASAN_ERRORS;
 use libafl::executors::inprocess::{HasInProcessHandlers, InProcessHandlers};
 
 /// The [`FridaInProcessExecutor`] is an [`Executor`] that executes the target in the same process, usinig [`frida`](https://frida.re/) for binary-only instrumentation.
-pub struct FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, S>
+pub struct FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, RT, S>
 where
     H: FnMut(&I) -> ExitKind,
     I: Input + HasTargetBytes,
@@ -31,12 +31,12 @@ where
     /// Frida's dynamic rewriting engine
     stalker: Stalker<'a>,
     /// User provided callback for instrumentation
-    helper: &'c mut FridaInstrumentationHelper<'b>,
+    helper: &'c mut FridaInstrumentationHelper<'b, RT>,
     followed: bool,
     _phantom: PhantomData<&'b u8>,
 }
 
-impl<'a, 'b, 'c, H, I, OT, S> Debug for FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, S>
+impl<'a, 'b, 'c, H, I, OT, RT, S> Debug for FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, RT, S>
 where
     H: FnMut(&I) -> ExitKind,
     I: Input + HasTargetBytes,
@@ -51,8 +51,8 @@ where
     }
 }
 
-impl<'a, 'b, 'c, EM, H, I, OT, S, Z> Executor<EM, I, S, Z>
-    for FridaInProcessExecutor<'a, 'b, 'c,  H, I, OT, S>
+impl<'a, 'b, 'c, EM, H, I, OT, RT, S, Z> Executor<EM, I, S, Z>
+    for FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, RT, S>
 where
     H: FnMut(&I) -> ExitKind,
     I: Input + HasTargetBytes,
@@ -93,8 +93,8 @@ where
     }
 }
 
-impl<'a, 'b, 'c, H, I, OT, S> HasObservers<I, OT, S>
-    for FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, S>
+impl<'a, 'b, 'c, H, I, OT, RT, S> HasObservers<I, OT, S>
+    for FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, RT, S>
 where
     H: FnMut(&I) -> ExitKind,
     I: Input + HasTargetBytes,
@@ -111,14 +111,18 @@ where
     }
 }
 
-impl<'a, 'b, 'c, H, I, OT, S> FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, S>
+impl<'a, 'b, 'c, H, I, OT, S, RT> FridaInProcessExecutor<'a, 'b, 'c, H, I, OT, RT, S>
 where
     H: FnMut(&I) -> ExitKind,
     I: Input + HasTargetBytes,
     OT: ObserversTuple<I, S>,
 {
     /// Creates a new [`FridaInProcessExecutor`]
-    pub fn new(gum: &'a Gum, base: InProcessExecutor<'a, H, I, OT, S>, helper: &'c mut FridaInstrumentationHelper<'b>) -> Self {
+    pub fn new(
+        gum: &'a Gum,
+        base: InProcessExecutor<'a, H, I, OT, S>,
+        helper: &'c mut FridaInstrumentationHelper<'b, RT>,
+    ) -> Self {
         let mut stalker = Stalker::new(gum);
         // Include the current module (the fuzzer) in stalked ranges. We clone the ranges so that
         // we don't add it to the INSTRUMENTED ranges.
